1.Creating a separated branch for each lesson, and organizing files per branch. 2.Calculator class fixed with double type instead of int type.
