+++ .gitignore
@@ -36,3 +36,11 @@ build/
 
 ### Mac OS ###
 .DS_Store
+
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
+hs_err_pid*
+replay_pid*
+.idea
+*.iml
+*.class
+
