+++ src/main/java/lesson2/task3/Calculator.java
@@ -1,0 +1,21 @@
+package lesson2.task3;
+
+public class Calculator {
+
+    public static double sum(double numberA, double numberB) {
+        return numberA + numberB;
+    }
+
+    public static double subtract(double numberA, double numberB) {
+        return numberA - numberB;
+    }
+
+    public static double multiply(double numberA, double numberB) {
+        return numberA * numberB;
+    }
+
+    public static double divide(double numberA, double numberB) {
+        return numberA / numberB;
+    }
+
+}
