--- src/main/java/org/example/AgeVerifier.java
@@ -1,11 +1,0 @@
-package org.example;
-
-public class AgeVerifier {
-    public static boolean verifyAgeGreaterOrEqual18(int age){
-        if(age >= 18){
-            return true;
-        }else{
-            return false;
-        }
-    }
-}
