--- src/main/java/org/example/Calculator.java
@@ -1,21 +1,0 @@
-package org.example;
-
-public class Calculator {
-
-    public static int sum(int numberA, int numberB) {
-        return numberA + numberB;
-    }
-
-    public static int subtract(int numberA, int numberB) {
-        return numberA - numberB;
-    }
-
-    public static int multiply(int numberA, int numberB) {
-        return numberA * numberB;
-    }
-
-    public static int divide(int numberA, int numberB) {
-        return numberA / numberB;
-    }
-
-}
