+++ src/test/java/lesson2/task3/CalculatorTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package lesson2.task3;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -10,53 +10,49 @@ class CalculatorTest {
     @Test
     void shouldAddTwoNumbersProperly() {
         //given
-        int numberA = 1;
-        int numberB = 2;
-        int expectedResult = 3;
+        double numberA = 1;
+        double numberB = 2;
+        double expectedResult = 3;
         //when
-        int result = Calculator.sum(numberA, numberB);
+        double result = Calculator.sum(numberA, numberB);
         //then
         Assertions.assertEquals(expectedResult, result);
-
     }
 
     @Test
     void shouldSubtractTwoNumbersProperly() {
         //given
-        int numberA = 2;
-        int numberB = 1;
-        int expectedResult = 1;
+        double numberA = 2;
+        double numberB = 1;
+        double expectedResult = 1;
         //when
-        int result = Calculator.subtract(numberA, numberB);
+        double result = Calculator.subtract(numberA, numberB);
         //then
         Assertions.assertEquals(expectedResult, result);
-
     }
 
     @Test
     void shouldMultiplyTwoNumbersProperly() {
         //given
-        int numberA = 2;
-        int numberB = 3;
-        int expectedResult = 6;
+        double numberA = 2;
+        double numberB = 3;
+        double expectedResult = 6;
         //when
-        int result = Calculator.multiply(numberA, numberB);
+        double result = Calculator.multiply(numberA, numberB);
         //then
         Assertions.assertEquals(expectedResult, result);
-
     }
 
     @Test
     void shouldDivideTwoNumbersProperly() {
         //given
-        int numberA = 10;
-        int numberB = 2;
-        int expectedResult = 5;
+        double numberA = 10;
+        double numberB = 2;
+        double expectedResult = 5;
         //when
-        int result = Calculator.divide(numberA, numberB);
+        double result = Calculator.divide(numberA, numberB);
         //then
         Assertions.assertEquals(expectedResult, result);
-
     }
 
 }
